Make spinlock errors more descriptive

+++ Common/Sync/kern_spinlock.hpp
@@ -17,7 +17,7 @@ public:
 	void Release();
 	BOOL TryAcquire();
 
-	[[nodiscard]] UINT8 Holder() const;
+	[[nodiscard]] INT16 Holder() const;
 	[[nodiscard]] BOOL IsLocked() const;
 
 	void SetDebugName(const char *Name);
@@ -26,7 +26,7 @@ public:
 private:
 	[[nodiscard]] BOOL IsHolding() const;
 	const char *DebugName;
-	UINT16 CoreNo;
+	INT16 CoreNo;
 	BOOL Locked; 
 };
